Rename links.ts to telephone-games.ts
