+++ stress-test/src/index.ts
@@ -85,6 +85,6 @@ const M = parseInt(process.argv[3], 10) || 1
 console.log(`Running stress tests with N=${N}, M=${M}`)
 
 require('./all-on')(orchestrator.registerScenario, N, M)
-require('./links')(orchestrator.registerScenario, N, M)
+require('./telephone-games')(orchestrator.registerScenario, N, M)
 
 orchestrator.run()
