+++ stress-test/src/telephone-games.ts
